12-10-23 IPC
Good news, basic SPI function seems to be good - the on-off segment functions well and the uPD7225 blink function works also.  There are problems with how the IC900 controller manages most of it's blink operations that likely will require a complete re-design of how the graphics are managed.  This shouldn't be a problem when used with my IC-900 controller clone, so victory is nigh! (neegh!).